Zoom out map to make NZ visible
With the google maps at zoom level 2 and
the chosen origin New Zealand is not visible
on the google map without zooming or scrolling.

Here we take the easy solution and zoom out to
level 1. We need to move the origin to the
equator so that we don't start with a grey bar
at the top of the map.

This isn't fantastic, as we now have lots of the
map dedicated to high latitudes, but should help
avoid us showing up on http://worldmapswithout.nz/